+++ resources/views/livewire/admin/user/includes/user-actions.blade.php
@@ -1,15 +1,19 @@
 <div class="flex items-center gap-2">
-    <x-shared.clickable
-        wire:loading.class='pointer-events-none animation-pulse'
-        wire:click="openUserFormModal('dialog_update_show', {{ $item->id }})"
-        icon="pencil"
-        small />
+    @can('update', $item)
+        <x-shared.clickable
+            wire:loading.class='pointer-events-none animation-pulse'
+            wire:click="openUserFormModal('dialog_update_show', {{ $item->id }})"
+            icon="pencil"
+            small />
+    @endcan
 
-    <x-shared.clickable
-        wire:loading.class='pointer-events-none animation-pulse'
-        wire:click="deleteUserConfirmation({{ $item->id }})"
-        style="danger"
-        variant="filled"
-        icon="trash"
-        small />
+    @can('delete', $item)
+        <x-shared.clickable
+            wire:loading.class='pointer-events-none animation-pulse'
+            wire:click="deleteUserConfirmation({{ $item->id }})"
+            style="danger"
+            variant="filled"
+            icon="trash"
+            small />
+    @endcan
 </div>
